pyproject.toml
[project]
name = "meru"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ftfy>=6.0",
    "future==0.18.0",
    "hydra-core>=1.1",
    "loguru>=0.6",
    "matplotlib>=3.10.3",
    "nltk>=3.6",
    "scikit-learn>=1.0",
    "seaborn>=0.13.2",
    "tensorboard==2.14.0",
    "tensorflow-datasets>=4.9.0",
    "timm>=1.0.15",
    "torchmetrics>=1.0",
    "torchvision>=0.22.1",
    "tqdm>=4.50.0",
    "ujson>=5.0",
    "umap-learn>=0.5.7",
    "webdataset>=0.2",
    "wordsegment>=1.3",
]

[tool.setuptools.packages.find]
include = ["meru*"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
